+++ settings.gradle
@@ -27,8 +27,6 @@ include ':simplified-books-registry-api'
 include ':simplified-boot-api'
 include ':simplified-bugsnag'
 include ':simplified-cardcreator'
-include ':simplified-datepicker'
-include ':simplified-datepicker-demo'
 include ':simplified-documents'
 include ':simplified-downloader-core'
 include ':simplified-feeds-api'
@@ -69,6 +67,8 @@ include ':simplified-tests-sandbox'
 include ':simplified-tests-strings'
 include ':simplified-threads'
 include ':simplified-ui-branding'
+include ':simplified-ui-datepicker'
+include ':simplified-ui-datepicker-demo'
 include ':simplified-ui-splash'
 include ':simplified-ui-theme'
