+++ simplified-app-shared/build.gradle
@@ -48,7 +48,6 @@ dependencies {
   api project(":simplified-boot-api")
   api project(":simplified-bugsnag")
   api project(":simplified-cardcreator")
-  api project(":simplified-datepicker")
   api project(":simplified-documents")
   api project(":simplified-downloader-core")
   api project(":simplified-feeds-api")
@@ -76,6 +75,7 @@ dependencies {
   api project(":simplified-tenprint")
   api project(":simplified-threads")
   api project(":simplified-ui-branding")
+  api project(":simplified-ui-datepicker")
   api project(":simplified-ui-splash")
   api project(":simplified-ui-theme")
