+++ simplified-app-vanilla-with-profiles/version.properties
@@ -1,4 +1,4 @@
 #
-#Thu Oct 10 10:35:10 UTC 2019
+#Thu Oct 10 11:14:46 UTC 2019
 versionName=3.0.14
-versionCode=4167
+versionCode=4174
