--- simplified-datepicker-demo/build.gradle
@@ -1,3 +1,0 @@
-dependencies {
-  implementation project(":simplified-datepicker")
-}
