--- simplified-datepicker-demo/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=org.librarysimplified.datepicker.demo
-POM_DESCRIPTION=Library Simplified (Date picker UI component demo)
-POM_NAME=org.librarysimplified.datepicker.demo
-POM_PACKAGING=apk
