--- simplified-datepicker/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=org.librarysimplified.datepicker
-POM_DESCRIPTION=Library Simplified (Date picker UI component)
-POM_NAME=org.librarysimplified.datepicker
-POM_PACKAGING=aar
