+++ simplified-tests/build.gradle
@@ -28,7 +28,6 @@ dependencies {
     exclude group: "com.github.tony19", module: "logback-android-classic"
   }
 
-  api project(":simplified-datepicker")
   api project(":simplified-documents")
   api project(":simplified-downloader-core")
   api project(":simplified-feeds-api")
@@ -56,6 +55,7 @@ dependencies {
   api project(":simplified-tests-strings")
   api project(":simplified-threads")
   api project(":simplified-ui-branding")
+  api project(":simplified-ui-datepicker")
   api project(":simplified-ui-splash")
   api project(":simplified-ui-theme")
