+++ simplified-ui-datepicker-demo/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+  implementation project(":simplified-ui-datepicker")
+}
