+++ simplified-ui-datepicker-demo/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=org.librarysimplified.ui.datepicker.demo
+POM_DESCRIPTION=Library Simplified (Date picker UI component demo)
+POM_NAME=org.librarysimplified.ui.datepicker.demo
+POM_PACKAGING=apk
