+++ simplified-ui-datepicker-demo/src/main/AndroidManifest.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.nypl.simplified.datepicker.demo">
+    package="org.nypl.simplified.ui.datepicker.demo">
 
   <application
       android:allowBackup="true">
 
     <activity
-        android:name="org.nypl.simplified.datepicker.demo.DatePickerDemo"
+        android:name="org.nypl.simplified.datepicker.ui.demo.DatePickerDemo"
         android:configChanges="orientation|screenSize">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
