+++ simplified-ui-datepicker/build.gradle
