+++ simplified-ui-datepicker/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=org.librarysimplified.ui.datepicker
+POM_DESCRIPTION=Library Simplified (Date picker UI component)
+POM_NAME=org.librarysimplified.ui.datepicker
+POM_PACKAGING=aar
