+++ simplified-ui-datepicker/src/main/res/layout/date_picker.xml
